Move funcs and consts to utils file
